requirements.txt
PyYAML==3.10
argparse==1.2.1
datadiff==1.1.6
dnspython==1.12.0
enum==0.4.4
hexdump==3.2
netaddr==0.7.14
python-magic==0.4.6
scapy==2.3.1
tld==0.7.2
scapy-http==1.6